fix(form): Avoid mutating Kamelet's schema
After updating the way how the Kamelet's model is updated, the `labels`
and `annotations` objects are mutated, which is not ideal since makes
the form complicated to use as the object fields are populated with the
standalone fields as well.

In addition to that, the Kamelet's custom schema is using Capitalized
version of the `type` entry, for instance, `Action`, `Source` and `Sink`

The fix is to recreate the object reference on each update and to fix
the schema enum by using the right values.

fix: https://github.com/KaotoIO/kaoto-next/issues/782